+++ src/main/java/com.accenture.flowershop/be/business/FlowerBusinessService.java
@@ -2,6 +2,7 @@ package com.accenture.flowershop.be.business;
 
 import com.accenture.flowershop.be.entity.flower.Flower;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface FlowerBusinessService {
@@ -10,5 +11,5 @@ public interface FlowerBusinessService {
     Flower findFlowerById(int idFlower);
     void deleteFlower(int idFlower);
     List<Flower> findFlowerByName(String nameFlower);
-    List<Flower> findFlowerByRangePrice(double minPrice, double maxPrice);
+    List<Flower> findFlowerByRangePrice(BigDecimal minPrice, BigDecimal maxPrice);
 }
